+++ src/utils.js
@@ -1,0 +1,5 @@
+const getRandomNumber = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min;
+const getRandomElement = (array) => array[Math.floor(Math.random() * array.length)];
+
+export {getRandomNumber};
+export {getRandomElement};
